Make order of Verilog files deterministic
Previously, the order of Verilog files was determined by `find` (in
order of directory entries), which varied from system to system
(depending on file creation/modification times). The order of Verilog
files matters, and there was a failure in CI when biriscv_defs.v
appeared after biriscv_fetch.v. This patch fixes the order to be
deterministic and adds this ordering requirement.